+++ data-redis-message/src/main/java/com/example/demo/DataInitializer.java
@@ -5,22 +5,15 @@
  */
 package com.example.demo;
 
-import java.nio.ByteBuffer;
-import java.util.UUID;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.ContextRefreshedEvent;
 import org.springframework.context.event.EventListener;
-import org.springframework.data.redis.connection.ReactiveKeyCommands;
-import org.springframework.data.redis.connection.ReactiveRedisConnection;
-import org.springframework.data.redis.connection.ReactiveRedisConnectionFactory;
-import org.springframework.data.redis.serializer.RedisSerializer;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
+
+import java.nio.ByteBuffer;
+import java.util.UUID;
 
 /**
  * @author hantsy
@@ -31,7 +24,6 @@ import reactor.core.publisher.Flux;
 class DataInitializer {
 
     private final PostRepository posts;
-
 
     @EventListener(value = ContextRefreshedEvent.class)
     public void init() {
@@ -44,15 +36,14 @@ class DataInitializer {
 
     private void initPosts() {
         this.posts.deleteAll()
-            .thenMany(
-                Flux.just("Post one", "Post two")
-                    .flatMap(title -> this.posts.save(Post.builder().id(UUID.randomUUID().toString()).title(title).content("content of " + title).build()))
-            );
-            //.subscribe(null, null, () -> log.info("done posts initialization..."));
+                .thenMany(
+                        Flux.just("Post one", "Post two")
+                                .flatMap(title -> this.posts.save(Post.builder().id(UUID.randomUUID().toString()).title(title).content("content of " + title).build()))
+                )
+                .subscribe(null, null, () -> log.info("done posts initialization..."));
     }
 
     private static String toString(ByteBuffer byteBuffer) {
-
         byte[] bytes = new byte[byteBuffer.remaining()];
         byteBuffer.get(bytes);
         return new String(bytes);
