+++ data-redis-message/src/main/java/com/example/demo/RedisConfig.java
@@ -5,36 +5,20 @@
  */
 package com.example.demo;
 
-import javax.annotation.PreDestroy;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
-import org.reactivestreams.Publisher;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.DataType;
 import org.springframework.data.redis.connection.ReactiveRedisConnectionFactory;
-import org.springframework.data.redis.connection.ReactiveSubscription;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.core.*;
-import org.springframework.data.redis.core.script.RedisScript;
+import org.springframework.data.redis.core.ReactiveRedisTemplate;
 import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.data.redis.listener.ReactiveRedisMessageListenerContainer;
-import org.springframework.data.redis.listener.Topic;
-import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
 import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.RedisElementReader;
-import org.springframework.data.redis.serializer.RedisElementWriter;
 import org.springframework.data.redis.serializer.RedisSerializationContext;
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import java.io.IOException;
-import java.time.Duration;
-import java.time.Instant;
-import java.util.List;
 import java.util.UUID;
 
 /**
@@ -65,8 +49,8 @@ public class RedisConfig {
     @Bean
     public ReactiveRedisTemplate<String, Post> reactiveRedisTemplate(ReactiveRedisConnectionFactory factory) {
         return new ReactiveRedisTemplate<String, Post>(
-            factory,
-            RedisSerializationContext.fromSerializer(new Jackson2JsonRedisSerializer(Post.class))
+                factory,
+                RedisSerializationContext.fromSerializer(new Jackson2JsonRedisSerializer(Post.class))
         );
     }
 
@@ -75,19 +59,19 @@ public class RedisConfig {
         ReactiveRedisMessageListenerContainer container = new ReactiveRedisMessageListenerContainer(connectionFactory);
         ObjectMapper objectMapper = new ObjectMapper();
         container.receive(ChannelTopic.of("posts"))
-            .map(p->p.getMessage())
-            .map(m -> {
-                try {
-                    Post post= objectMapper.readValue(m, Post.class);
-                    post.setId(UUID.randomUUID().toString());
-                    return post;
-                } catch (IOException e) {
-                    return null;
-                }
-            })
-            .switchIfEmpty(Mono.error(new IllegalArgumentException()))
-            .flatMap(p-> posts.save(p))
-            .subscribe(c-> log.info(" count:" + c), null , () -> log.info("saving post."));
+                .map(p -> p.getMessage())
+                .map(m -> {
+                    try {
+                        Post post = objectMapper.readValue(m, Post.class);
+                        post.setId(UUID.randomUUID().toString());
+                        return post;
+                    } catch (IOException e) {
+                        return null;
+                    }
+                })
+                .switchIfEmpty(Mono.error(new IllegalArgumentException()))
+                .flatMap(p -> posts.save(p))
+                .subscribe(c -> log.info(" count:" + c), null, () -> log.info("saving post."));
         return container;
     }
