+++ data-redis/pom.xml
@@ -21,9 +21,7 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
-		<spring-data-releasetrain.version>Lovelace-RC2</spring-data-releasetrain.version>
-		<reactor-bom.version>Californium-M2</reactor-bom.version>
+        <maven.compiler.release>17</maven.compiler.release>
     </properties>
 
     <dependencies>    
@@ -86,10 +84,10 @@
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-classic</artifactId>
         </dependency>
-        
+
         <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
@@ -107,15 +105,15 @@
             <artifactId>reactor-test</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>org.springframework.security</groupId>
-            <artifactId>spring-security-test</artifactId>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
 
     <build>
         <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.1</version>
+            </plugin>
             <!-- Maven Assembly Plugin -->
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
@@ -143,7 +141,6 @@
 				</executions>
 			</plugin>
 
-            
         </plugins>
     </build>
 
@@ -184,6 +181,5 @@
             </snapshots>
         </pluginRepository>
     </pluginRepositories>
-
 
 </project>
