+++ data-redis/src/main/java/com/example/demo/RedisConfig.java
@@ -5,31 +5,13 @@
  */
 package com.example.demo;
 
-import javax.annotation.PreDestroy;
-
-import org.reactivestreams.Publisher;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.DataType;
 import org.springframework.data.redis.connection.ReactiveRedisConnectionFactory;
-import org.springframework.data.redis.connection.ReactiveSubscription;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.core.*;
-import org.springframework.data.redis.core.script.RedisScript;
-import org.springframework.data.redis.listener.Topic;
-import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
+import org.springframework.data.redis.core.ReactiveRedisTemplate;
 import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.RedisElementReader;
-import org.springframework.data.redis.serializer.RedisElementWriter;
 import org.springframework.data.redis.serializer.RedisSerializationContext;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
-
-import java.time.Duration;
-import java.time.Instant;
-import java.util.List;
 
 /**
  * @author hantsy
@@ -58,8 +40,8 @@ public class RedisConfig {
     @Bean
     public ReactiveRedisTemplate<String, Post> reactiveRedisTemplate(ReactiveRedisConnectionFactory factory) {
         return new ReactiveRedisTemplate<String, Post>(
-            factory,
-            RedisSerializationContext.fromSerializer(new Jackson2JsonRedisSerializer(Post.class))
+                factory,
+                RedisSerializationContext.fromSerializer(new Jackson2JsonRedisSerializer(Post.class))
         );
     }
